MBS-3416: bucket now generates a unique ID that is deterministic and based on the inputs
It is cleaner and better to generate an id from bucket contents instead of allowing user to specify the id manually via constructor.
In latter case it is possible to use the same ID for entirely different buckets, and that is an error prone.
Now we hash bucket contents and generate ID from that.