MBS-4162: support --test-arg-file
In order to have a flexible way of scheduling tests to be run, we allow to pass a JSON file with
a description of all tests that must be run. This allows to specify a various combinations of test+test destination,
instead of executing a matrix of test x test destinations.