MBS-4272: report bucketIdsBeingProcessed
When worker fetches a bucket, and then some local error occurs when it attempts to run tests from that bucket,
the result will not be reported back, it will be considered as a lost bucket. In this case, queue server will still
wait for result to come back.
In order to make sure the worker processing state and queue server expectation are in sync, we now pass all
buckets that worker is currently processing via REST call (`ReportAliveRequest`). This will allow queue to
understand when worker has dropped bucket and re-enqueue it back to the queue.